Remove Program Counter table and instead incement Program Counter with memory read and writes.
Turns out this was a dumb idea as it resulted in my jump instructions always being off by usually +3. Normally the PG is incremented during the instruction which would then be overwritten by the jump instruction. But I incremented after jumps so you never ended up where you inteded. I considered just manually changing the jump instructions to take care of it but that would result in instructions that doesn't match online explanations and would just generally be logically weird. Instead I'll increment the PG when reading and writing to memory.